# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "salus",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sharing/leaderboard": {
      "get": {
        "summary": "Api Leaderboard List",
        "operationId": "api_leaderboard_list_api_v1_sharing_leaderboard_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Api Leaderboard Create",
        "operationId": "api_leaderboard_create_api_v1_sharing_leaderboard_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_CreateLeaderboardBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/leaderboard/{group_id}": {
      "get": {
        "summary": "Api Leaderboard Get",
        "operationId": "api_leaderboard_get_api_v1_sharing_leaderboard__group_id__get",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Api Leaderboard Delete",
        "operationId": "api_leaderboard_delete_api_v1_sharing_leaderboard__group_id__delete",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/leaderboard/{group_id}/join": {
      "post": {
        "summary": "Api Leaderboard Join",
        "operationId": "api_leaderboard_join_api_v1_sharing_leaderboard__group_id__join_post",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_JoinLeaderboardBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/leaderboard/{group_id}/leave": {
      "post": {
        "summary": "Api Leaderboard Leave",
        "operationId": "api_leaderboard_leave_api_v1_sharing_leaderboard__group_id__leave_post",
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/connections": {
      "get": {
        "summary": "Api Connections List",
        "operationId": "api_connections_list_api_v1_sharing_connections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Api Connections Create",
        "operationId": "api_connections_create_api_v1_sharing_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_CreateConnectionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/connections/{connection_id}/accept": {
      "post": {
        "summary": "Api Connections Accept",
        "operationId": "api_connections_accept_api_v1_sharing_connections__connection_id__accept_post",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/connections/{connection_id}/decline": {
      "post": {
        "summary": "Api Connections Decline",
        "operationId": "api_connections_decline_api_v1_sharing_connections__connection_id__decline_post",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sharing/connections/{connection_id}": {
      "delete": {
        "summary": "Api Connections Delete",
        "operationId": "api_connections_delete_api_v1_sharing_connections__connection_id__delete",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Connection Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/federation/sharing": {
      "get": {
        "summary": "Federated Shared Data",
        "operationId": "federated_shared_data_api_v1_federation_sharing_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "owner_username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Owner Username"
            }
          },
          {
            "name": "data_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Data Type"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "_CreateConnectionBody": {
        "properties": {
          "grantee_handle": {
            "type": "string",
            "title": "Grantee Handle"
          },
          "metric_code": {
            "type": "string",
            "title": "Metric Code"
          },
          "aggregation_level": {
            "type": "string",
            "title": "Aggregation Level",
            "default": "daily_summary"
          }
        },
        "type": "object",
        "required": [
          "grantee_handle",
          "metric_code"
        ],
        "title": "_CreateConnectionBody"
      },
      "_CreateLeaderboardBody": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "metric_type_code": {
            "type": "string",
            "title": "Metric Type Code",
            "default": "steps"
          },
          "time_frame": {
            "type": "string",
            "title": "Time Frame",
            "default": "weekly"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "_CreateLeaderboardBody"
      },
      "_JoinLeaderboardBody": {
        "properties": {
          "invite_code": {
            "type": "string",
            "title": "Invite Code"
          }
        },
        "type": "object",
        "required": [
          "invite_code"
        ],
        "title": "_JoinLeaderboardBody"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}